feat!: extend int, int8, int16, int32, int64, uint, uint8, uint16, uint32, uint64 and custom func convert.
BREAKING CHANGE: `Int` struct retype to `Number` struct, add Type field to find out what type in use.
enhanced `FromInterface` function now support more type like `Struct` `Interface` `Func`.

+++ eval_test.go
@@ -21,17 +21,17 @@ func TestEval(t *testing.T) {
 	}
 
 	eval(`undefined`, nil, nil)
-	eval(`1`, nil, int64(1))
-	eval(`19 + 23`, nil, int64(42))
+	eval(`1`, nil, 1)
+	eval(`19 + 23`, nil, 42)
 	eval(`"foo bar"`, nil, "foo bar")
-	eval(`[1, 2, 3][1]`, nil, int64(2))
+	eval(`[1, 2, 3][1]`, nil, 2)
 
 	eval(
 		`5 + p`,
 		map[string]interface{}{
 			"p": 7,
 		},
-		int64(12),
+		12,
 	)
 	eval(
 		`"seven is " + p`,
